% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angulo.R
\name{angulo}
\alias{angulo}
\title{Internal. Angle associated to the spatial gradient}
\usage{
angulo(dx, dy)
}
\arguments{
\item{dx}{\code{numeric} giving the longitudinal gradient component}

\item{dy}{\code{numeric} giving the latitudinal gradient component
angulo()}
}
\description{
Internal. Angle associated to the spatial gradient
}
