% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anSeries.R
\name{anSeries}
\alias{anSeries}
\title{Summary annual values for the climatic series.}
\usage{
anSeries(r, yr1, yr2, yr0, comp = "mean")
}
\arguments{
\item{r}{\code{RasterStack} with mean monthly values of the climatic variable
for the period of interest. Only series with complete years (no missing data)
should be used.}

\item{yr0, 1, 2}{\code{Integers} specifying the first (yr0) year in the series, 
the initial (yr1) and the last (yr2) years for the period of interest.}

\item{comp}{\code{logical} specifying the summary statistic to be computed. 
Must be one of mean (default), max or min.}
}
\value{
A \code{RasterStack} with the annual mean, max, or min values.
}
\description{
Function to calculate the mean/max/min monthly annual series from a monthly  
series for a period of interest.
}
\author{
Jorge Garcia Molinos
}
