% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcslope.R
\name{calcslope}
\alias{calcslope}
\title{Calculate trend in temperature for pixels in a raster brick}
\usage{
calcslope(x, divisor = 10)
}
\arguments{
\item{rx}{\code{brick} of rasters giving temperature at years.}

\item{divisor}{\code{numeric} giving divisor for rate.}
}
\value{
A \code{data.frame} giving the slope at each pixel and number of
points in each pixel.
}
\description{
Calculate trend in temperature for pixels in a raster brick
}
\details{
Calculates the trend in temperature.
\code{divisor} is for getting correct units, e.g. if data is annual then
\code{divisor = 10} gives degrees C per decade.
This method is pretty fast. It
takes about 2.5 seconds on a 3Ghz laptop to proces 64800 cells over
80 years of annual data.
}
\examples{
data(sst)
dat <- calcslope(sst)
head(dat$slope)
}
\author{
David Schoeman, Christopher Brown
}
