% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climPlot.R
\name{climPlot}
\alias{climPlot}
\title{Binned scatter plot for 2-dimensional climate space}
\usage{
climPlot(xy, x.binSize, y.binSize)
}
\arguments{
\item{xy}{\code{data.frame} with cells as rows and 4 columns representing the present and future local values for the two variables (V1p, V1f, V2p, V2f).}

\item{x.binSize}{\code{numeric} the bin size for the first variable.}

\item{y.binSize}{\code{numeric} the bin size for the second variable.}
}
\value{
a series of \code{plot} objects displaying the (i) present and (ii) future
cell frequency for each combination of local climates,
and (iii) the location of remnant, novel and disappearing climates between both periods.
}
\description{
Function to create a binned scatter plot of two climate variables.
}
\examples{

data(JapTC)
# Plot climate space for the two first variables(annual precipitation and maximum temperature)
xy <- na.omit(data.frame(getValues(JapTC[[1]]), getValues(JapTC[[2]]),
getValues(JapTC[[3]]), getValues(JapTC[[4]])))
climPlot(xy, x.binSize = 5, y.binSize = 0.2)

}
\seealso{
{\code{\link{climAna}}, \code{\link{climPCA}}}
}
\author{
Jorge Garcia Molinos
}
