% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EEZ}
\alias{EEZ}
\title{Icelandic Exclusive Economic Zone (EEZ)}
\format{
A spatial polygon data frame containing the Icenalndic EEZ
(200 NM).
}
\source{
Marineregions.org
\url{http://www.marineregions.org/downloads.php}
}
\description{
Icelandic Exclusive Economic Zone (v10).
}
