% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEZs.R
\docType{data}
\name{EEZs}
\alias{EEZs}
\title{Exclusive Economic Zone Boundaries (EEZ, v9)}
\format{A spatial polygon data frame containing 10 arbitrarily selected EEZs.}
\source{
Marineregions.org
\url{http://www.marineregions.org/downloads.php}
}
\usage{
EEZs
}
\description{
Exclusive Economic Zone Boundaries (EEZ, v9)
}
\keyword{datasets}
