% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gVoCC.R
\name{gVoCC}
\alias{gVoCC}
\title{Gradient-based climate velocity}
\usage{
gVoCC(tempTrend, spatGrad)
}
\arguments{
\item{tempTrend}{The output from the tempTrend function containing the long-term linear climatic trends.}

\item{spatGrad}{The output from the spatGrad function containing the magnitudes and angles for the spatial climatic gradient.}
}
\value{
A \code{RasterStack} containing the climate velocity magnitude ("voccMag",
km/yr for unprojected rasters and spatial unit/year for projected rasters) and angle ("voccAng" in
degrees north: 0N, 90E, 180S and 270W).
}
\description{
Function to calculate the velocity of climate change after Burrows et al. (2011)
based on local climatic temporal trends and spatial gradients.
}
\examples{

HSST <- terra::rast(system.file("extdata", "HadiSST.tif", package = "VoCCdata"))

yrSST <- sumSeries(HSST,
  p = "1960-01/2009-12", yr0 = "1955-01-01", l = terra::nlyr(HSST),
  fun = function(x) colMeans(x, na.rm = TRUE), freqin = "months", freqout = "years"
)
tr <- tempTrend(yrSST, th = 10)
sg <- spatGrad(yrSST, th = 0.0001, projected = FALSE)

# Magnitude and angle of the climate velocity (km/yr) 1960-2009

v <- gVoCC(tr, sg)
terra::plot(v)
}
\references{
\href{http://science.sciencemag.org/content/334/6056/652}{Burrows et al. 2011}. The pace of shifting climate
in marine and terrestrial ecosystems. Science, 334, 652-655.
}
\seealso{
\code{\link{tempTrend}}, \code{\link{spatGrad}}
}
