% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JapTC.R
\docType{data}
\name{JapTC}
\alias{JapTC}
\title{Multivariate Terraclimate data for Japan}
\format{A raster stack with 9 layers for the historical (1960-1969) and current (2008-2017)
average (AnMn) and standard deviation (AnMnSD) of the mean annual (AnMn) precipitation (Ppr),
maximum (Tmax) and minimum (Tmin) temperature.}
\source{
\href{http://www.climatologylab.org/terraclimate.html}{Terraclimate} data set.
}
\usage{
data(JapTC)
}
\description{
Historical (1960-1969) and present (2008-2017) multivariate 1/24
Terraclimate data set (Abatzoglou et al. 2018), extracted for the Japanese
archipelago and including 3 variables: mean annual monthly precipitation (mm),
maximum and minimum temperatures (C).
}
\references{
\href{https://www.nature.com/articles/sdata2017191}{Abatzoglou et al. 2018}. Terraclimate, a high-resolution global
dataset of monthly climate and climatic water balance from 1958-2015, Scientific Data, 5: 170191
}
\keyword{datasets}
