% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lVoCC.R
\name{lVoCC}
\alias{lVoCC}
\title{Local climate velocity}
\usage{
lVoCC(tempTrend, spatGrad)
}
\arguments{
\item{tempTrend}{The output from the tempTrend function containing the long-term linear climatic trends.}

\item{spatGrad}{The output from the tempTrend function containing the magnitudes and angles for the spatial climatic gradient.}
}
\value{
A \code{RasterStack} containing the climate velocity magnitude ("voccMag",  
km/yr for unprojected rasters and spatial unit/year for projected rasters) and
angle("voccAng" in degrees C).
}
\description{
Function to calculate the velocity of climate change after Burrows et al. 2011 
based on local climatic temporal trends and spatial gradients.
}
\author{
Jorge Garcia Molinos
}
