% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajLine.R
\name{trajLine}
\alias{trajLine}
\title{Trajectory spatial lines}
\usage{
trajLine(x)
}
\arguments{
\item{x}{\code{data.frame} containing the coordinates (x, y) of the constituent
points and identification number (trajIDs) for each trajectory as returned by VoCCTraj.}
}
\value{
A \code{SpatialLinesDataFrame} with one line per trajectory as specified in x.
To avoid artifacts, trajectories crossing the date line need to be splitted into two segments.
Where the trajectory on one side of the date line is only composed of a single point, 
the trajectory won't be displayed (no line object created). The function assumes 
a -180 to 180 longitudinal arrengment.
}
\description{
Create a spatial line data frame object from trajectory points.
}
\author{
Jorge Garcia Molinos
}
