% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocc.R
\docType{package}
\name{vocc}
\alias{vocc}
\alias{vocc-package}
\title{Tools for calculating the Velocity of Climate Change from
raster data}
\description{
The \pkg{vocc} provides functions for calculating the VoCC.
This is a beta-version
}
\details{
Try \code{vignette("vocc")} for more guidance.
See
\code{\link{spatialgrad}} for calculating spatial gradients.
\code{\link{calcslope}} for getting temperature trends.
\code{\link{calcvelocity}} for calculating velocity trends.
}
