% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{VoCC_get_data}
\alias{VoCC_get_data}
\title{Return data location and data if requested}
\usage{
VoCC_get_data(path = NULL)
}
\arguments{
\item{path}{Filename of data if known.}
}
\value{
The path as a character string, or a dataset
}
\description{
Return data location and data if requested
}
\examples{

VoCC_get_data()

HSST <- VoCC_get_data(path = "HSST")

}
