% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voccSST.R
\docType{data}
\name{voccSST}
\alias{voccSST}
\title{Average (1960-2009) mean monthly sea surface temperatures and associated climate velocity}
\format{A raster stack containing the grand mean SST, and the magnitude and angle 
of the climate velocity.}
\source{
Mean monthly SSTs for the period 1960-2009 from the HSST data and
associated velocity of climate change.
}
\usage{
voccSST
}
\description{
Average (1960-2009) mean monthly sea surface temperatures and associated climate velocity
}
\keyword{datasets}
