% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angulo.R
\name{angulo}
\alias{angulo}
\title{Internal. Angle associated to the spatial gradient}
\usage{
angulo(dx, dy)
}
\description{
Internal. Angle associated to the spatial gradient
}
\author{
Jorge Garcia Molinos and David S. Schoeman
angulo()
}
