% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcvelocity.R
\name{calcvelocity}
\alias{calcvelocity}
\title{Calculate velocity from the spatial gradient and temporal trends}
\usage{
calcvelocity(grad, slope, y_dist = 111.325)
}
\arguments{
\item{grad}{\code{data.frame} returned from \code{spatialgrad}}

\item{slope}{\code{data.frame} returned from \code{calcslope}}

\item{y_dist}{\code{numeric} giving units (degrees latitude) for the coordinate}
}
\value{
\code{data.frame} giving velocity.
}
\description{
Calculate velocity from the spatial gradient and temporal trends
}
\examples{
data(sst)
slopedat <- calcslope(sst)
allyears <- rep(1, raster::nlayers(sst))
mnsst <- raster::stackApply(sst, indices = allyears, fun = mean)
spatx <- spatialgrad(mnsst)
velodf <- calcvelocity(spatx, slopedat)
r <- raster::raster(sst)
r[velodf$icell] <- velodf$velocity
raster::plot(r)
}
\author{
Christopher J. Brown
}
