% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEZ.R
\docType{data}
\name{EEZ}
\alias{EEZ}
\title{Icelandic Exclusive Economic Zone (EEZ)}
\format{A spatial polygon data frame containing the Icenalndic EEZ
(200 NM).}
\source{
Marineregions.org
\url{http://www.marineregions.org/downloads.php}
}
\usage{
data(EEZ)
}
\description{
Icelandic Exclusive Economic Zone (v10).
}
\keyword{datasets}
