% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEZs.R
\docType{data}
\name{EEZs}
\alias{EEZs}
\title{Exclusive Economic Zone Boundaries (EEZ, v10)}
\format{A spatial polygon data frame containing EEZs for
Western Pacific tropical island nations}
\source{
Marineregions.org
\url{http://www.marineregions.org/downloads.php}
}
\usage{
data(EEZs)
}
\description{
Western Pacific tropical Exclusive Economic Zones
}
\keyword{datasets}
