% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSST.R
\docType{data}
\name{HSST}
\alias{HSST}
\title{Global monthly mean sea surface temperatures (SST) for 1955-2010}
\format{A raster stack with 672 layers containing global mean monthly SSTs (C)
at 1-deg resolution for the period Jan 1955 to Dec 2010.}
\source{
Hadley Centre data set \href{https://www.metoffice.gov.uk/hadobs/hadisst/}{HadISST 1.1} (data accessed May 2018).
}
\usage{
data(HSST)
}
\description{
Global monthly mean sea surface temperatures (SST) for 1955-2010
}
\references{
\href{https://www.metoffice.gov.uk/hadobs/hadisst/HadISST_paper.pdf}{Rayner et al. 2003}. Global analyses of sea
surface temperature, sea ice, and night marine air temperature since the late nineteenth century. J. Geophys. Res.Vol. 108: 4407.
}
\keyword{datasets}
