% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSST.R
\docType{data}
\name{HSST}
\alias{HSST}
\title{Global monthly mean sea surface temperatures (SST) for 1960-2009}
\format{A raster stack with 672 layers containing global mean monthly SSTs at 1-deg resolution
for the period Jan 1955 to Dec 2010.}
\source{
Hadley Centre data set HadISST 1.1 (Rayner et al. 2003)
\url{https://www.metoffice.gov.uk/hadobs/hadisst/}
}
\usage{
HSST
}
\description{
Global monthly mean sea surface temperatures (SST) for 1960-2009
}
\keyword{datasets}
