% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resTime.R
\name{resTime}
\alias{resTime}
\title{Climatic residence time of a polygon}
\usage{
resTime(pg, vel, areapg = NA)
}
\arguments{
\item{pg}{\code{SpatialPolygon} or a \code{SpatialPolygonsDataFrame} containing the polygons for which
the residence time is to be calculated. The polygons must be on the same coordinate system as vel.}

\item{vel}{\code{raster} with climate velocity (km/year) for the period of interest.}

\item{areapg}{\code{vector} with the area (in km2) of the polygons. Use NA (default) to calculate internally if field not avilable.}
}
\value{
a \code{data.frame} containing for each polygon its ID, mean velocity (km/yr),
diameter of the equivalent circle (km), and residence time (years) as the ratio D/vel.
}
\description{
Function to calculate vocc residence time of a polygon after Loaire et al. (2009)
}
\examples{

# Load example Exclusive Economic Zone polygons

?EEZs
?HSST
yrSST <- sumSeries(HSST, p = "1969-01/2009-12", yr0 = "1955-01-01", l = nlayers(HSST),
fun = function(x) colMeans(x, na.rm = TRUE),
freqin = "months", freqout = "years")
tr <- tempTrend(yrSST, th = 10)
sg <- spatGrad(yrSST, th = 0.0001, projected = FALSE)
v <- gVoCC(tr,sg)
vel <- v[[1]]
# Calculating area internally
a1 <- resTime(EEZs, vel, areapg = NA)
a1
# Using the area field from the polygon data table
a2 <- resTime(EEZs, vel, areapg = as.numeric(as.numeric(levels(EEZs$Area_km2))[EEZs$Area_km2]))
a2
}
\references{
\href{https://www.nature.com/articles/nature08649}{Loarie et al. 2009}. The velocity of climate change. Nature, 462, 1052-1055.
}
\seealso{
{\code{\link{gVoCC}}}
}
\author{
Jorge Garcia Molinos
}
