% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftTime.R
\name{shiftTime}
\alias{shiftTime}
\title{Shift in timing of seasonal climatology}
\usage{
shiftTime(r, yr1, yr2, yr0, th, month)
}
\arguments{
\item{r}{\code{stack} with monthly values of the climatic
variable for the period of interest.}

\item{yr1}{\code{integer} specifying the initial year for the period of interest.}

\item{yr2}{\code{integer} specifying the end year for the period of interest.}

\item{yr0}{\code{integer} specifying the first year in the series.}

\item{th}{\code{integer} max number of non NAs in the series needed to
calculate the trend (default 3).}

\item{m}{\code{integer} number (1-12) of the month for which the shift is to be calculated}
}
\value{
a \code{stack} with the long-term monthly trend (C/year; "mTrend"),
seasonal rate of change (C/month; "seaRate"), and seasonal shift (day/decade; "seaShift").
}
\description{
Function to calculate the seasonal shift in the arriving of seasonal climatology
for a given period of interest as per Burrows et al. 2011.
}
\examples{
data(HSST)
Apr <- shiftTime(HSST, yr1 = 1960, yr2 = 2009, yr0 = 1955, th = 10, m = 4)
plot(Apr)

}
\references{
\href{http://science.sciencemag.org/content/334/6056/652}{Burrows et al. 2011}. The pace of shifting climate in marine and terrestrial ecosystems. Science, 334, 652-655.
}
\author{
Jorge Garcia Molinos and Michael T. Burrows
}
