% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatGrad.R
\name{spatGrad}
\alias{spatGrad}
\title{Local spatial climatic gradients}
\usage{
spatGrad(r, th = -Inf, projected = FALSE)
}
\arguments{
\item{r}{\code{RasterStack} with the annual climatic values for the period of interest.
Alternatively, a \code{raster} with the annual climatic values averaged
over the period of interest.}

\item{th}{\code{Integer} indicating a lower thershold to truncate the spatial
gradient with. Use -Inf (default) if no threshold required.}

\item{projected}{\code{Logical} is the source raster in a projected coordinate system?
If FALSE (default) a correction will be made to account for latitudinal distortion.}
}
\value{
A \code{RasterStack} with the magnitude of the spatial gradient
(Grad in C per km for unprojected rasters and C per spatial unit for projected rasters),
and the associated angle (Ang in degrees).
}
\description{
Function to calculate the magnitude and direction of the spatial gradient
associated to a climatic variable after Burrows et al. (2011). This trend is
to be used for the calculation of the gradient-based climate velocity using gVoCC.
}
\examples{

yrSST <- sumSeries(HSST, p = "1969-01/2009-12", yr0 = "1955-01-01", l = nlayers(HSST),
fun = function(x) colMeans(x, na.rm = TRUE), freqin = "months", freqout = "years")

# Spatial gradient (magnitude and angle) for the average mean annual SST.

sg <- spatGrad(yrSST, th = 0.0001, projected = FALSE)

plot(sg)

}
\references{
\href{http://science.sciencemag.org/content/334/6056/652}{Burrows et al. 2011}. The pace of shifting climate in marine and terrestrial ecosystems. Science, 334, 652-655.
}
\seealso{
{\code{\link{tempTrend}}, \code{\link{gVoCC}}}
}
\author{
Jorge Garcia Molinos, David S. Schoeman, and Michael T. Burrows
}
