% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialgrad.R
\name{spatialgrad}
\alias{spatialgrad}
\title{Calculate spatial gradient across a raster layer}
\usage{
spatialgrad(rx, y_dist = c(111.325, 111.325), y_diff = 1)
}
\arguments{
\item{rx}{\code{raster} giving raster to calculate spatial gradient on}

\item{y_dist}{\code{numeric} giving distance for WE and NS gradient for
1 unit of the y-coordinate}

\item{y_diff}{\code{numeric} giving units (degrees latitude) for the pixel
dimensions}
}
\value{
\code{data.frame} giving \code{icell} raster cell numbers,
  \code{WE} west-east gradient, \code{NS} north-south gradient,
  \code{angle} the angle of the trajectory.
}
\description{
Calculate spatial gradient across a raster layer
}
\details{
Assumes unprojected coordinates (ie lon-lat).
If you are using projected coordinates, with distance in metres,
you should give \code{y_dist = res(rx)} and \code{y_diff = NA}.
If \code{y_diff = NA}, no correction will be made for changing
size of 1 degree latitude with latitude.
}
\examples{
data(sst)
allyears <- rep(1, raster::nlayers(sst))
mnsst <- raster::stackApply(sst, indices = allyears, fun = mean)
spatx <- spatialgrad(mnsst)
head(spatx$NS)
}
\author{
Christopher J. Brown
}
