% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitLine.R
\name{splitLine}
\alias{splitLine}
\title{Internal. Split a line segment defined by points A-B into n parts}
\usage{
splitLine(A, B, n)
}
\description{
Internal. Split a line segment defined by points A-B into n parts
}
\author{
Jorge Garcia Molinos
}
