% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempTrend.R
\name{tempTrend}
\alias{tempTrend}
\title{Long-term local climatic trends}
\usage{
tempTrend(r, th)
}
\arguments{
\item{r}{\code{RasterStack} with annual values of the climatic variable for
the period of interest.}

\item{th}{\code{Integer} max number of observations in the series needed to
calculate the trend at each cell.}
}
\value{
A \code{RasterStack} containing annual trends
("slpTrends" in degree celsius per year) and their associated SE ("seTrends")
and statistical significance("sigTrends").
}
\description{
Function to calculate temporal trend from the mean annual raster series
of a climatic variable. This trend is to be used for the calculation of the
local climate velocity using lVoCC.
}
\examples{

data(HSST)
yrSST <- sumSeries(HSST, p = "1969-01/2009-12", yr0 = "1955-01-01", l = nlayers(HSST),
fun = function(x) colMeans(x, na.rm = TRUE), freqin = "months", freqout = "years")

# Mean annual SST trend (minimum threshold of 10 years of data), with SE and p-values.

tr <- tempTrend(yrSST, th = 10)

plot(tr)

}
\author{
Jorge Garcia Molinos and Christopher J. Brown
}
