% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj25.R
\docType{data}
\name{traj25}
\alias{traj25}
\title{Climate velocity trajectories}
\format{A data frame containing coordinates and id for each trajectory.}
\source{
climate velocity trajectories produced using the HSST data (mean monthly sea surface temperatures) 
for a 0.25-degree starting coordinate grid and the period 1960-2009.
}
\usage{
traj25
}
\description{
Climate velocity trajectories
}
\keyword{datasets}
