% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voccTraj.R
\name{voccTraj}
\alias{voccTraj}
\title{Climate velocity trajectories}
\usage{
voccTraj(lonlat, vel, ang, mn, tyr, trajID = 1:nrow(lonlat))
}
\arguments{
\item{lonlat}{\code{data.frame} with the longitude and latitude (in decimal degrees)
of the points to project.}

\item{vel}{\code{raster} with the magnitude of local climate velocity.}

\item{ang}{\code{raster} with velocity angles.}

\item{mn}{\code{raster} with the overal mean annual climatic value over the period of interest.}

\item{tyr}{\code{integer} number of years comprising the period of interest.}
}
\value{
a \code{data.frame} containing the coordinates ("x", "y") of the constituent
points and identification number ("trajIDs") for each trajectory.
}
\description{
Function to calculate vocc trajectories after Burrows et al (2014). Trajectories
are calculated by propagating climatic isopleths using the magnitude and direction of 
local (cell) velocities. This is a slightly modified version of the original Burrows et al. (2014)
approach in that iterations of a trajectory are based on cumulative time travelled instead of using fixed time steps.
}
\author{
Jorge Garcia Molinos and David S. Schoeman
}
